[FEAT/#14] OtherUserChatProfile 분리
